fix thank-you slide bug
Adding a thank-you title slide at the end of the presentation resulted
in title slides not having distinct styles. This was caused by the fact
that all content under title slides is simply discarded, and a script
necessary for title slide styling to work was the last thing in the body
of the presentation (i.e. it was under the thank-you slide if that was
included, and was being discarded).

Instead, the javascript snippet has been moved into an included header,
which fixes the problem, is cleaner and also less prone to breakage due
to inadvertent deletion of the snippet while replacing the contents of
the presentation.